Getting closer to fixing Windows import.
By looking at platform.py in Blender\pythin\lib, it seems that the best
function to use to use the patch on Linux and not Windows is if
platform.system_alias(). However, 'pass' skips the rest of the import
process, and I keep getting 'continue of of loop' error. Is there
something I'm missing?