Can I please undo all the horrible changes I did
in the last few days? Thank you. From now on, big changes such as the
Linux patch and material nodes will be done on the non-master branch.

Reverted back to V0.8 Beta 1, but with a notable change: I added a small
part of Marc's patch, and now the LDraw path can be set from within
Blender! :D