Detection no longer overwrites previous path, but a new problem as emerged.
If the path contains \t, \n, \r, or the like, it is interpreted rather
than treated like a raw string, as it should be (example:
C:\temp\LDraw).  I added 'r' to each string and converted all single
backslashes to os.path.sep, but it still doesn't work. Also, the `if
"\\" in` lines are untested, I added them right now. That may have fixed
it, but I do not know yet.