This is the start of displaying the LEGO text on studs.
The intention of this 5 month old code, applied only to the v2 code base, was to detect if the currently file was one of the stud parts, and if so replace it with a stud part that had the LEGO text. I do not not recall why I did not publish this, but I do believe it would have worked once it was figured out what file to use and if it should be bundeled with the script.