Restore debugPrint() for patch compatibility
This lead way for multiple changes to the Console class:
+ Add new private __makeMessage() method to construct the message
+ Add new warn() static method
+ Update log() method to use __makeMessage()
+ Better message formatting
+ debugPrint() displays a deprecation message when called

The behavior of debugPrint has been altered slightly, though it was
never fully documented to begin with. Originally it worked in the same
manner as print(), accepting multiple unnamed parameters and
constructing one message. Now, it only accepts a single string (just as
it was already used) and the Console methods accept the old behavior.